[MINOR]: rename file to optional_u128
